Clean up nested path example
There is some semantic confusion due to the use of `event` for both the event system and the thing which are listed in the calendar. I renamed it to `meeting` where applicable to resolve this.

Also, I think I fixed a problem in the second part of the example where there were still references to the indices and `event`.

Anyways, it looks like events are deprecated in favor of channels, but for now I thought I'd fix the example.